+++ nym-vpn-desktop/src/constants.ts
@@ -1,18 +1,5 @@
 import { Country } from './types';
 
-export const routes = {
-  root: '/',
-  settings: '/settings',
-  display: '/settings/display',
-  logs: '/settings/logs',
-  feedback: '/settings/feedback',
-  feedbackSend: '/settings/feedback/send',
-  support: '/settings/support',
-  legal: '/settings/legal',
-  entryNodeLocation: '/entry-node-location',
-  exitNodeLocation: '/exit-node-location',
-} as const;
-
 export const AppName = 'NymVPN';
 export const ConnectionEvent = 'connection-state';
 export const ProgressEvent = 'connection-progress';
