+++ nym-vpn-desktop/src/pages/index.ts
@@ -2,4 +2,5 @@ export * from './home';
 export * from './settings';
 export { default as MainLayout } from './MainLayout';
 export * from './location';
+export * from './login';
 export { default as Error } from './Error';
