+++ nym-vpn-desktop/src/pages/login/index.ts
@@ -1,0 +1,1 @@
+export { default as LogIn } from './LogIn';
