+++ nym-vpn-desktop/src/types/routes.ts
@@ -1,3 +1,3 @@
-import { routes } from '../constants';
+import { routes } from '../router';
 
 export type Routes = (typeof routes)[keyof typeof routes];
